Add compatibility for Lootr.
This patch simply switches to the LockableLootTileEntity's static
`setLootTable` method instead of directly accessing the tile entity and
calling it there.

This provides compatibility with Lootr, a mod that provides instanced
loot, per-chest, per-player, for multiplayer servers. Unfortunately, I
can't hook into the member `setLootTable` as the world will either be
null or not be the relevant world-generation world.

Switching to the static method means the world-generation world instance
is passed in and that can be hooked into to provide replacements without
causing crashes or deadlocks.